backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
sqlalchemy[asyncio]>=2.0.36
asyncpg>=0.30.0
aiosqlite>=0.20.0
pydantic>=2.10.0
pydantic-settings>=2.6.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.17
alembic>=1.14.0
aiosmtplib>=3.0.2
email-validator>=2.2.0
slowapi>=0.1.9
bcrypt>=4.2.0
bleach>=6.1.0
